% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{DistConect}
\alias{DistConect}
\title{Calculates the proportion of a species that could migrate from a cell to its
neighbours}
\usage{
DistConect(Raster, Distance, Time = 7)
}
\arguments{
\item{Raster}{a raster with the space where the species will be inhabiting}

\item{Distance}{the maximum dispersal distance of the species}

\item{Time}{the number of time-slices to be used}
}
\value{
a dataframe with the cell id from, the cell id to, and beta, that is
the proportion of individuals that will go from cell from to cell to
}
\description{
Calculates the proportion of a species that could migrate from a cell to its
neighbours given a raster, the maximum dispersal distance, and the number of
times the time-lapse will have
}
\examples{
data("r")
DistConect(Raster = r, Distance = 1000000, Time = 7)
}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>

Javier Fajardo <javierfajnolla@gmail.com >
}
