% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{MultiSppQuad}
\alias{MultiSppQuad}
\title{Generates an AMPL dat file from a list of Stacks}
\usage{
MultiSppQuad(Stacklist, Dist, name, nchains = 100, costlayer)
}
\arguments{
\item{Stacklist}{a list of Stacks with the space where the species will be
inhabiting}

\item{Dist}{the maximum dispersal distance of the species modeled in the
stack}

\item{name}{the name of the .dat file that will be exported}

\item{nchains}{the number of chains to go through}

\item{costlayer}{raster with the costs of each cell}
}
\value{
exports a .dat file to feed the AMPL model
}
\description{
Generates an AMPL dat file from a Stack in which each file is the projection
of a species distribution model into a time slice
}
\examples{
data("BinSpp")
MultiSppQuad(Stacklist = BinSpp, Dist = 1000000, name = "Two")
}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>

Javier Fajardo <javierfajnolla@gmail.com >
}
