% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{RasterToAmplDat}
\alias{RasterToAmplDat}
\title{Generates an AMPL dat file from a Stack}
\usage{
RasterToAmplDat(Stack, maxalpha = 10, maxbiomass = 2, maxcapacidad = 10,
  name = "Stack", Dist = 1e+06)
}
\arguments{
\item{Stack}{a raster with the space where the species will be inhabiting}

\item{maxalpha}{the maximum rate of change of the population in optimal
coditions, defaults in 10}

\item{maxbiomass}{the maximum initial biomass of the population in optimal
coditions, defaults in 2}

\item{maxcapacidad}{the maximum biomass of the population in optimal coditions,
defaults in 10}

\item{name}{the name of the .dat file that will be exported, defaults in
stack}

\item{Dist}{the maximum dispersal distance of the species modeled in the
stack}
}
\value{
exports a .dat file to feed the AMPL model
}
\description{
Generates an AMPL dat file from a Stack in which each file is the projection
of a species distribution model into a time slice
}
\examples{
data("univariate")
RasterToAmplDat(Stack = univariate)
}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>
}
