% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{RtoQuadAmplDat}
\alias{RtoQuadAmplDat}
\title{Generates an AMPL dat file from a Stack}
\usage{
RtoQuadAmplDat(Stack, Distance, Threshold, name)
}
\arguments{
\item{Stack}{a raster with the space where the species will be inhabiting}

\item{Distance}{the maximum dispersal distance of the species modeled in the
stack}

\item{Threshold}{minimum value in the model to allow the species to exist}

\item{name}{the name of the .dat file that will be exported, defaults in
stack}

\item{Time}{number of time slices present in the Stack}
}
\value{
exports a .dat file to feed the AMPL model
}
\description{
Generates an AMPL dat file from a Stack in which each file is the projection
of a species distribution model into a time slice
}
\examples{
data("univariate")
RtoQuadAmplDat(Stack = univariate, Distance = 1000000, Threshold = 0.5,
name = "TeSt")
}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>

Javier Fajardo <javierfajnolla@gmail.com >
}
